+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ### Added
-- An overlapping utility library.
+- An lexical overlapping utility module (`overlap`).
+- A token normalizer that merges tokens in to spans (`JoinTokenMapper`).
 
 
 ## [1.0.1] - 2022-01-25